Controller for homepage
It contains the function required to get data from form and save it in
the database.

Bill amount is calculated using the requirements of previous assignment.

One flash message will be shown to user for the final bill amount. User
has to select atlas one topping to put his order.